refactor: 디렉토리 이름 변경
