+++ src/main/java/com/example/javaproject2/week7/day30_31/model/Person.java
@@ -1,4 +1,4 @@
-package com.example.javaproject2.week7.day30.model;
+package com.example.javaproject2.week7.day30_31.model;
 
 public class Person {
     private int id;
